+++ sample/Source/SentryPlayground.Target.cs
@@ -8,15 +8,12 @@ public class SentryPlaygroundTarget : TargetRules
 	public SentryPlaygroundTarget( TargetInfo Target) : base(Target)
 	{
 		Type = TargetType.Game;
-		DefaultBuildSettings = BuildSettingsVersion.V2;
+		DefaultBuildSettings = BuildSettingsVersion.Latest;
+
+#if UE_5_1_OR_LATER
+		IncludeOrderVersion = EngineIncludeOrderVersion.Latest;
+#endif
+
 		ExtraModuleNames.AddRange( new string[] { "SentryPlayground" } );
-
-		bOverrideBuildEnvironment = true;
-
-		if (Target.Version.MajorVersion == 5 && Target.Version.MinorVersion < 2)
-		{
-			bUsePCHFiles = false;
-			bUseUnityBuild = false;
-		}
 	}
 }