+++ sample/Source/SentryPlayground/SentryPlaygroundUtils.cpp
@@ -2,10 +2,7 @@
 
 void USentryPlaygroundUtils::Crash()
 {
-	// Supressing warnings for when using warnings as errors on the target. 
-	#pragma warning(suppress: 6011)	
-	char *ptr = 0;
-	#pragma warning(suppress: 6011)
+	volatile char *ptr = 0;
 	*ptr += 1;
 }